+++ src/main/java/com/trevorism/https/DefaultSecureHttpClient.java
@@ -1,5 +1,6 @@
 package com.trevorism.https;
 
+import com.trevorism.http.HttpClient;
 import com.trevorism.https.token.ObtainTokenFromAuthServiceFromPropertiesFile;
 
 public class DefaultSecureHttpClient extends SecureHttpClientBase{
@@ -7,4 +8,5 @@ public class DefaultSecureHttpClient extends SecureHttpClientBase{
     public DefaultSecureHttpClient() {
         super(new ObtainTokenFromAuthServiceFromPropertiesFile());
     }
+
 }