+++ src/main/java/com/trevorism/https/SecureHttpClient.java
@@ -1,16 +1,17 @@
 package com.trevorism.https;
 
-public interface SecureHttpClient {
+import com.trevorism.http.HttpClient;
+
+public interface SecureHttpClient extends HttpClient {
     String AUTHORIZATION = "Authorization";
     String BEARER_ = "bearer ";
-
+    String CORRELATION_ID_HEADER_KEY = "X-Correlation-ID";
     /**
      * Performs an HTTP GET
      *
      * @param url The url to GET
      * @return The string content of the response
      */
-    String get(String url);
     String get(String url, String correlationId);
 
     /**
@@ -20,7 +21,6 @@ public interface SecureHttpClient {
      * @param serialized The serialized item to POST
      * @return The string content of the response
      */
-    String post(String url, String serialized);
     String post(String url, String serialized, String correlationId);
 
     /**
@@ -30,8 +30,16 @@ public interface SecureHttpClient {
      * @param serialized The serialized item to PUT
      * @return The string content of the response
      */
-    String put(String url, String serialized);
     String put(String url, String serialized, String correlationId);
+
+    /**
+     * Performs an HTTP PATCH
+     *
+     * @param url        The url to PATCH
+     * @param serialized The serialized item to PATCH
+     * @return The string content of the response
+     */
+    String patch(String url, String serialized, String correlationId);
 
     /**
      * Performs an HTTP DELETE
@@ -39,6 +47,5 @@ public interface SecureHttpClient {
      * @param url The url to DELETE
      * @return The string content of the response
      */
-    String delete(String url);
     String delete(String url, String correlationId);
 }