+++ src/main/java/com/trevorism/https/token/InvalidTokenCredentialsException.java
@@ -4,4 +4,7 @@ public class InvalidTokenCredentialsException extends RuntimeException {
     public InvalidTokenCredentialsException() {
         super("Bad Request to get token");
     }
+    public InvalidTokenCredentialsException(Throwable e) {
+        super("Bad Request to get token", e);
+    }
 }