+++ src/main/java/com/trevorism/https/token/ObtainTokenFromAuthServiceFromPropertiesFile.java
@@ -1,7 +1,8 @@
 package com.trevorism.https.token;
 
-import com.trevorism.secure.ClasspathBasedPropertiesProvider;
-import com.trevorism.secure.PropertiesProvider;
+import com.trevorism.ClasspathBasedPropertiesProvider;
+import com.trevorism.PropertiesProvider;
+import com.trevorism.http.HttpClient;
 
 public class ObtainTokenFromAuthServiceFromPropertiesFile extends ObtainTokenFromAuthService{