+++ src/main/java/com/trevorism/https/token/ObtainTokenFromParameter.java
@@ -1,5 +1,7 @@
 package com.trevorism.https.token;
 
+import com.trevorism.http.HttpClient;
+
 public class ObtainTokenFromParameter implements ObtainTokenStrategy{
 
     private final String token;
@@ -7,6 +9,11 @@ public class ObtainTokenFromParameter implements ObtainTokenStrategy{
     @Override
     public String getToken() {
         return token;
+    }
+
+    @Override
+    public void setHttpClient(HttpClient client) {
+        //not needed, other implementations may need a HttpClient
     }
 
     public ObtainTokenFromParameter(String token){