+++ changelog.md
@@ -1,3 +1,11 @@
+# 2.0.0
+
+Upgrading to 5.0.0 secure utils, which removes Jersey specific code.
+
+Upgrading to 2.0.0 http utils
+
+Update the interface to allow SecureHttpClient to extend HttpClient
+
 # 1.0.0
 
 Upgrading to 4.0.0 secure utils, uses lazy loading of secrets.